requirements.txt
fastapi         # our web framework
uvicorn         # the server to run FastAPI
sqlalchemy      # the database ORM (object-relations mapping using Python instead of raw SQL)
psycopg2-binary # PostgreSQL driver
pydantic        # data validation
python-jose     # for JSON web token authentication
passlib         # for password hashing
celery          # for asynchronous tasks
redis           # the message broker for celery
pytest          # for unit testing
pytest-asyncio
httpx           # for testing FastAPI
alembic         # for tracking changes to database
email-validator # for EmailStr validation
bcrypt==4.0.1   # missing from passlib library in python 3.14
cryptography    # backup for passlib incase bcrypt doesn't work